Rename directory and add then statement
